requirements.txt
pocketflow
python-telegram-bot
google-generativeai
gspread
google-auth-oauthlib
python-dotenv
matplotlib